Change van_dam to manyfold, and update image path
I wasn't sure about changing the <name> field, I thought that might break existing installations, so I've left that for now. I'll create a copy of this XML in the manyfold3d organisation soon, but I thought a quick fix before I get to that would be a good idea.